+++ js/data/change-lives.js
@@ -3,9 +3,6 @@ export const changeLives = (game, lives) => {
     throw new Error(`Lives should be of type number`);
   }
 
-  if (lives < 0) {
-    throw new Error(`Lives should not be negative value`);
-  }
   const newGame = Object.assign({}, game, {
     lives
   });
