+++ js/data/quest.js
@@ -1,6 +1,6 @@
 export const INITIAL_GAME = Object.freeze({
   level: 0,
-  lives: 4,
+  lives: 3,
   time: 0,
   answers: []
 });
