--- js/modules/01-intro-screen.js
@@ -1,17 +1,0 @@
-import {getElementFromTemplate, changeScreen} from './util.js';
-import greetingScreen from './02-greeting-screen.js';
-
-const template = `<section class="intro">
-  <button class="intro__asterisk asterisk" type="button"><span class="visually-hidden">Продолжить</span>*</button>
-  <p class="intro__motto"><sup>*</sup> Это не фото. Это рисунок маслом нидерландского художника-фотореалиста Tjalf Sparnaay.</p>
-</section>`;
-
-
-const element = getElementFromTemplate(template);
-const agreeButton = element.querySelector(`button.intro__asterisk`);
-
-agreeButton.addEventListener(`click`, () => {
-  changeScreen(greetingScreen);
-});
-
-export default element;
