+++ js/modules/abstract-view.js
@@ -1,0 +1,32 @@
+const render = (string) => {
+  const el = document.createElement(`div`);
+  el.innerHTML = string.trim();
+  return el;
+};
+
+class AbstractView {
+  constructor() {
+    if (new.target === AbstractView) {
+      throw new Error(`Can't instantiate AbstractView, only concrete one`);
+    }
+  }
+  get template() {
+    throw new Error(`Template is required`);
+  }
+  get element() {
+    if (this._element) {
+      return this._element;
+    }
+    this._element = this.render();
+    this.bind(this._element);
+
+    return this._element;
+  }
+  render() {
+    return render(this.template);
+  }
+  bind() {
+  }
+}
+
+export default AbstractView;
