+++ js/modules/greeting/greeting-screen.js
@@ -1,0 +1,12 @@
+import {changeScreen} from './../util.js';
+import GreetingView from './greeting-view.js';
+import {getRulesScreen} from './../rules/rules-screen.js';
+
+export const getGreetingScreen = () => {
+  const template = new GreetingView();
+  template.onClick = () => {
+    changeScreen(getRulesScreen());
+  };
+
+  return template.element;
+};
