+++ js/modules/statistics/statistics-screen.js
@@ -1,0 +1,7 @@
+import StatisticsView from './statistics-view.js';
+
+export const getStatictics = (state) => {
+  const template = new StatisticsView(state);
+
+  return template.element;
+};
