+++ js/modules/status-bar/status-bar-screen.js
@@ -1,0 +1,7 @@
+import StatusBarView from './status-bar-view.js';
+
+export const getStatusBar = (state) => {
+  const template = new StatusBarView(state);
+
+  return template.element;
+};
